packages/pin-server/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
RPi.GPIO==0.7.0
Werkzeug==0.15.5
